move block preview logic
